Add image figure + ,clean code

+++ src/Controller/CreateController.php
@@ -12,7 +12,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
-
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 use Symfony\Component\Routing\Annotation\Route;
 
 
@@ -26,8 +26,9 @@ class CreateController extends AbstractController
     }
 
     #[Route('/add/figure', name: 'app_add_figure')]
-    public function index(TrickManager $manager, Request $request): Response
+    public function index(TrickManager $manager, Request $request,AuthorizationCheckerInterface $authorizationChecker): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $figure = new Figure();
 
         $form = $this->createForm(FigureFormType::class, $figure);
@@ -39,7 +40,7 @@ class CreateController extends AbstractController
             return $this->redirectToRoute('app_index');
         }
 
-        return $this->render('', [
+        return $this->render('add_figure/index.html.twig', [
             'form' => $form->createView(),
         ]);
     }
@@ -62,25 +63,40 @@ class CreateController extends AbstractController
     }
 
     #[Route('/figure/{id}/delete', name: 'figure_delete', methods: ['GET', 'POST'])]
-    public function delete(Figure $figure, EntityManagerInterface $entityManager, Request $request): Response
+    public function delete(Figure $figure, EntityManagerInterface $entityManager, Request $request, AuthorizationCheckerInterface $authorizationChecker): Response
     {
-//        $this->denyAccessUnlessGranted('CAN_DELETE', $figure);
+        if (!$authorizationChecker->isGranted('FIGURE_DELETE', $figure)) {
+            $this->addFlash('error', 'Vous ne pouvez pas supprimer cette figure.');
+            return $this->redirectToRoute('app_index');
+        }
 
         if ($this->isCsrfTokenValid('delete' . $figure->getId(), $request->request->get('_token'))) {
             $entityManager->remove($figure);
             $entityManager->flush();
+
+            $this->addFlash('success', 'Figure supprimée avec succès !');
+        } else {
+            $this->addFlash('error', 'Invalid CSRF token');
         }
 
         return $this->redirectToRoute('app_index');
     }
 
     #[Route('/figure/{id}/edit', name: 'figure_edit', methods: ['GET', 'POST'])]
-    public function edit(TrickManager $manager, Request $request, Figure $figure): Response {
+    public function edit(TrickManager $manager, Request $request, Figure $figure, AuthorizationCheckerInterface $authorizationChecker): Response
+    {
+        if (!$authorizationChecker->isGranted('FIGURE_EDIT', $figure)) {
+            $this->addFlash('error', 'Vous ne pouvez pas modifier cette figure.');
+            return $this->redirectToRoute('app_index');
+        }
+
         $form = $this->createForm(FigureFormType::class, $figure);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+
             $manager->update($form->get('image'), $figure);
+            $this->addFlash('success', 'Votre figure a bien été modifiée !');
             return $this->redirectToRoute('figure_show', ['id' => $figure->getId()]);
         }
